+++ testdata/apps/embedded/app/contatos/page.go
@@ -1,0 +1,5 @@
+package contatos
+
+func Page() {}
+
+func POST() {}