+++ testdata/apps/methodmw/app/api/funis/route.go
@@ -1,0 +1,5 @@
+package funis
+
+func GET() {}
+
+func POST() {}